Refactor review thumbnail grid per code review feedback.
Extract preview scheduler, thumbnail hook, and shared types; use a single resize message listener and JS-only minHeight for subgrid row growth while CSS handles visual scaling.
