+++ code/core/src/shared/constants/iframe-resize.ts
@@ -1,0 +1,2 @@
+/** postMessage `context` for embed iframe content dimension updates (preview → parent). */
+export const IFRAME_RESIZE_CONTEXT = 'iframe.resize';
